Update ncov-qc-runner.py
Added a lot of changes. The script now can create a directory where the input barcodes.csv file is located and write a config.yaml file to it. The config is fairly customisable, and can be adapted for illumina or nanopore sequencing outputs. Section to run ncov-tools from this script is currently not working with os.system commands.